Port to SDL 3
Also have stb_image use custom malloc

+++ Source/Headers/input.h
@@ -86,7 +86,7 @@ void InitInput(void);
 void ReadKeyboard(void);
 
 OGLPoint2D GetLogicalMouseCoord(void);
-void GetMousePixelCoord(int *x, int *y);
+void GetMousePixelCoord(float *x, float *y);
 
 void DoSDLMaintenance(void);
 Boolean GetKeyState(uint16_t sdlScancode);
